refactor: worker.tsのリファクタリングとテスト整備
## 主な変更

### worker.tsのファイル分割
- handleCardByIdとhandleStatsをhandlers/info.tsに移動
- worker.tsを純粋なルーティング層として整理（207行→159行）
- 全てのビジネスロジックをlib/worker/以下に統一

### テスト構成の整備
- tests/README.mdを作成してテスト構成を文書化
- handlers/info.tsのユニットテストを追加（10テスト）
- 既存テストを修正して実装をインポートするように変更
  - worker-api.test.ts: 実際のparseFilterParams等をインポート
  - worker.test.ts: 実際のparseSearchQuery等をインポート
  - handlers-info.test.ts: 実際のhandleCardById等をインポート

### 成果
- テスト総数: 235 → 245（全て成功）
- 実装変更時にテストが失敗するようになり、仕様の検証が可能に
- コード構造が整理され保守性が向上
