+++ src/__tests__/worker.test.ts
@@ -1,115 +1,10 @@
 /**
- * Tests for search-parser functions in worker.ts
+ * Tests for search-parser functions
  */
 
 import { describe, it, expect } from 'vitest'
-
-// Copy interfaces and functions from worker.ts for testing
-interface ParsedSearchQuery {
-  positive: string[]
-  negative: string[]
-  phrases: string[]
-  regexes: string[]
-}
-
-function parseSearchQuery(query: string): ParsedSearchQuery {
-  const result: ParsedSearchQuery = {
-    positive: [],
-    negative: [],
-    phrases: [],
-    regexes: []
-  }
-
-  if (!query || !query.trim()) {
-    return result
-  }
-
-  const matched: Array<{ start: number; end: number }> = []
-
-  // 1. Extract regex patterns
-  const regexPattern = /reg:(?:"([^"]+)"|(\S+))/g
-  let match: RegExpExecArray | null
-  while ((match = regexPattern.exec(query)) !== null) {
-    const pattern = match[1] || match[2]
-    result.regexes.push(pattern)
-    matched.push({ start: match.index, end: match.index + match[0].length })
-  }
-
-  // 2. Extract negative search
-  const negativePattern = /-(?:"([^"]+)"|(\S+))/g
-  while ((match = negativePattern.exec(query)) !== null) {
-    const word = match[1] || match[2]
-    result.negative.push(word)
-    matched.push({ start: match.index, end: match.index + match[0].length })
-  }
-
-  // 3. Extract phrase search
-  const phrasePattern = /"([^"]+)"/g
-  while ((match = phrasePattern.exec(query)) !== null) {
-    const isOverlapping = matched.some(m =>
-      (match!.index >= m.start && match!.index < m.end) ||
-      (match!.index + match![0].length > m.start && match!.index + match![0].length <= m.end)
-    )
-
-    if (!isOverlapping) {
-      const phrase = match[1]
-      result.phrases.push(phrase)
-      matched.push({ start: match.index, end: match.index + match[0].length })
-    }
-  }
-
-  // 4. Extract positive words
-  let remaining = ''
-  let lastEnd = 0
-
-  matched.sort((a, b) => a.start - b.start)
-
-  for (const m of matched) {
-    remaining += query.substring(lastEnd, m.start) + ' '
-    lastEnd = m.end
-  }
-  remaining += query.substring(lastEnd)
-
-  const words = remaining.split(/\s+/).filter(w => w.trim() && w !== '-' && !w.startsWith('reg:'))
-  result.positive = words
-
-  return result
-}
-
-function matchParsedQuery(text: string, parsed: ParsedSearchQuery): boolean {
-  if (!text) return false
-
-  for (const negWord of parsed.negative) {
-    if (text.includes(negWord)) {
-      return false
-    }
-  }
-
-  for (const regexStr of parsed.regexes) {
-    try {
-      const regex = new RegExp(regexStr)
-      if (!regex.test(text)) {
-        return false
-      }
-    } catch (e) {
-      return false
-    }
-  }
-
-  for (const phrase of parsed.phrases) {
-    if (!text.includes(phrase)) {
-      return false
-    }
-  }
-
-  for (const word of parsed.positive) {
-    if (!text.includes(word)) {
-      return false
-    }
-  }
-
-  return true
-}
+import { parseSearchQuery, matchParsedQuery } from '../lib/shared/search-parser.js'
+import type { ParsedSearchQuery } from '../lib/shared/search-parser.js'
 
 describe('parseSearchQuery', () => {
   it('should parse negative search', () => {
